+++ src/ducks/counter/counterSaga.js
@@ -1,15 +1,15 @@
 import { delay } from 'redux-saga'
-import { call, put, takeLatest } from 'redux-saga/effects'
-import counterActions, { INCREMENT_REQUESTED } from 'ducks/counter/counterActions.js';
+import { call, put, takeLatest } from 'redux-saga/effects';
+import { INCREMENT_REQUESTED, increment } from 'ducks/counter/counterActions';
+
 
 function* incrementAsyncWatcher () {
   yield takeLatest([INCREMENT_REQUESTED], incrementAsyncWorker)
 }
 
 export function* incrementAsyncWorker (action) {
-  yield call(delay, 1000)
-  yield put(counterActions.increment())
+  yield call(delay, 100)
+  yield put(increment())
 }
 
-export const sagas = [incrementAsyncWatcher]
-
+export const counterSagas = [incrementAsyncWatcher]
