+++ src/ducks/rootReducers.js
@@ -1,0 +1,7 @@
+import count from 'ducks/counter/counterReducer';
+
+const reducers = {
+  count
+}
+
+export default reducers;
