+++ src/ducks/rootSagas.js
@@ -1,17 +1,17 @@
 import { call, spawn } from 'redux-saga/effects';
-import { sagas as counterSagas } from 'ducks/count';
+import { counterSagas } from 'ducks/counter/counterSaga';
 import count from 'ducks/counter/counterReducer';
 
 const reducers = {
   count
 }
 
-export default reducers
+export default reducers;
 
 export function* rootSaga () {
-  const sagas = [...counterSagas]
+  const allCounterSagas = [...counterSagas];
 
-  yield sagas.map(saga =>
+  yield allCounterSagas.map(saga =>
     spawn(function* () {
       let isSyncError = false
       while (!isSyncError) {
@@ -27,5 +27,5 @@ export function* rootSaga () {
         }
       }
     })
-  )
+  );
 }
